Moved handle export csv function to a new file
